# Auth/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Тестовое задание от MEDODS",
        "title": "Сервис аутентификации",
        "contact": {
            "name": "Evans Trein",
            "url": "https://github.com/EvansTrein",
            "email": "evanstrein@icloud.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "paths": {
        "/auth/delSession/{guid}": {
            "delete": {
                "description": "Удаление сессии с указанным GUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Удаление сессии",
                "parameters": [
                    {
                        "type": "string",
                        "format": "7c5e66cf-57ba-4871-9186-74ff5ab1e1f1",
                        "description": "Уникальный идентификатор сессии",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponceMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Обновляет AccessToken и RefreshToken токены",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh operation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RefreshToken",
                        "name": "RefreshToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    }
                }
            }
        },
        "/auth/{guid}": {
            "post": {
                "description": "Аутентификация пользователя с предоставленным GUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "format": "7c5e66cf-57ba-4871-9186-74ff5ab1e1f1",
                        "description": "GUID пользователя",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для входа",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponce"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AuthData": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user1@mail.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                }
            }
        },
        "models.ErrResponce": {
            "type": "object",
            "properties": {
                "errMessage": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "models.ResponceMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "info message"
                }
            }
        },
        "models.Tokens": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJjbGllbnRJUCI6IjEyNy4wLjAuMSIsImV4cCI6MTczMzE4MzY5MSwicmVmcmVzaFRva2VuSUQiOjExfQ.LyuwUe7IPSG2_aPdT59Ms2_xmDPa9-ymhGsuuJ_uwi5wzxfjoHerNSTpJLf2ZQUXGNjDp3BHgs2jXw4ehLLjuQ"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "3r65EyQIo/NsGR3TE1/Y7GIuD+jm1diGf+zZ4DoXwhg="
                }
            }
        },
...
    },
...
}